fix tests for OOP
